refactor: change `plugin/sample` package to `sample`
To make the conventional `go build ./...` and `go install ./...`
work. With the sample in the `main` package currently:

```shellsession
$ go build ./...
runtime.main_main·f: function main is undeclared in the main package
```

Closes https://github.com/daveshanley/vacuum/issues/135
