+++ plugin/sample/check_single_path.go
@@ -1,4 +1,4 @@
-package main
+package sample
 
 import (
 	"fmt"
